Fixed some hanging notes bugs Increased precision to 32 bit (prevent clipping) Added new chorus Added phaser Fixed UI Added fixed mid-eq to better model the MKS-20 circuit

+++ librdpiano/include/sound_chip.h
@@ -11,7 +11,7 @@ public:
   u8 read(size_t offset);
   void write(size_t offset, u8 data);
 
-  s16 update();
+  s32 update();
 
   void load_samples(const u8 *temp_ic5, const u8 *temp_ic6, const u8 *temp_ic7);
